refactor: estimated returns UI for users with no positions
