+++ app/position-management/position-details.tsx
@@ -6,9 +6,17 @@ import { Progress } from '@/components/ui/progress'
 import { BodyText, HeadingText } from '@/components/ui/typography'
 import useGetPortfolioData from '@/hooks/useGetPortfolioData'
 import { useSearchParams } from 'next/navigation'
-import React, { useContext, useState } from 'react'
+import React, { useContext } from 'react'
 import { motion } from 'framer-motion'
-import { abbreviateNumber, capitalizeText, convertScientificToNormal, getLiquidationRisk, getLowestDisplayValue, getRiskFactor, hasLowestDisplayValuePrefix, isLowestValue } from '@/lib/utils'
+import {
+    abbreviateNumber,
+    capitalizeText,
+    convertScientificToNormal,
+    getLiquidationRisk,
+    getRiskFactor,
+    hasLowestDisplayValuePrefix,
+    isLowestValue
+} from '@/lib/utils'
 import { AssetsDataContext } from '@/context/data-provider'
 import AvatarCircles from '@/components/ui/avatar-circles'
 import useGetPlatformData from '@/hooks/useGetPlatformData'
@@ -19,32 +27,14 @@ import Image from 'next/image'
 import { useActiveAccount, useConnect } from 'thirdweb/react'
 import TooltipText from '@/components/tooltips/TooltipText'
 import InfoTooltip from '@/components/tooltips/InfoTooltip'
-import {
-    Card,
-    CardContent,
-    CardHeader,
-} from "@/components/ui/card"
-import { Slider } from "@/components/ui/slider"
-import { TPlatform } from '@/types'
-
-type TRow = {
-    id: number;
-    key: "lend" | "borrow" | "duration";
-    title: string;
-    logo?: string;
-    selectedLabel: string;
-    selectedValue: number;
-    totalValue: number;
-    step: number;
-}
+import { EstimatedReturns } from './estimated-returns'
+import { getStatDisplayValue } from './helper-functions'
 
 export default function PositionDetails() {
     const searchParams = useSearchParams();
     const { allChainsData } = useContext(AssetsDataContext);
-    // const tokenAddress = searchParams.get("token") || "";
     const chain_id = searchParams.get("chain_id") || 0;
     const protocol_identifier = searchParams.get("protocol_identifier") || "";
-    // const { address: walletAddress, isConnecting, isDisconnected } = useAccount();
     const activeAccount = useActiveAccount();
     const walletAddress = activeAccount?.address;
     const { isConnecting } = useConnect();
@@ -142,7 +132,7 @@ export default function PositionDetails() {
     // Loading state
     if (isLoading) {
         return (
-            <div className="w-full h-[150px] rounded-6 overflow-hidden">
+            <div className="w-full h-[200px] rounded-6 overflow-hidden">
                 <Skeleton className='w-full h-full' />
             </div>
         )
@@ -166,7 +156,7 @@ export default function PositionDetails() {
     // If user is connected, but does not have any positions, show estimated returns
     if (!isLoading && userPositions.length === 0) {
         return (
-            <EsimatedReturns
+            <EstimatedReturns
                 platformDetails={platformData}
             />
         )
@@ -288,173 +278,4 @@ export default function PositionDetails() {
             </div>
         </motion.section>
     )
-}
-
-// Child components
-function EsimatedReturns({
-    platformDetails
-}: {
-    platformDetails: TPlatform;
-}) {
-    const [selectedValue, setSelectedValue] = useState({
-        lend: 0,
-        borrow: 0,
-        duration: 0,
-    });
-
-    const lendAssetDetails = platformDetails?.assets.filter(asset => !asset.borrow_enabled)[0];
-    const borrowAssetDetails = platformDetails?.assets.filter(asset => asset.borrow_enabled)[0];
-
-    const supplyAPY = lendAssetDetails?.supply_apy;
-    const borrowAPY = borrowAssetDetails?.variable_borrow_apy;
-    const amountSupplied = selectedValue.lend;
-    const amountBorrowed = selectedValue.borrow;
-    const supplyTokenPrice = supplyAPY * lendAssetDetails?.token?.price_usd;
-    const borrowTokenPrice = borrowAPY * borrowAssetDetails?.token?.price_usd;
-    const duration = selectedValue.duration;
-
-    const handleSelectedValueChange = (value: number, type: "lend" | "borrow" | "duration") => {
-        setSelectedValue(prev => ({ ...prev, [type]: value }));
-    }
-
-    const rows: TRow[] = [
-        {
-            id: 1,
-            key: "lend",
-            title: "lend collateral",
-            logo: lendAssetDetails?.token.logo,
-            selectedLabel: lendAssetDetails?.token.symbol,
-            selectedValue: selectedValue.lend,
-            totalValue: 50,
-            step: 1,
-        },
-        {
-            id: 2,
-            key: "borrow",
-            title: "borrowing",
-            logo: borrowAssetDetails?.token.logo,
-            selectedLabel: borrowAssetDetails?.token.symbol,
-            selectedValue: selectedValue.borrow,
-            totalValue: 50,
-            step: 1,
-        },
-        {
-            id: 3,
-            key: "duration",
-            title: "Duration",
-            selectedLabel: "years",
-            selectedValue: selectedValue.duration,
-            totalValue: 5,
-            step: 0.5,
-        }
-    ];
-
-    const estimatedEarnings = getEstimatedEarnings({
-        supplyAPY,
-        borrowAPY,
-        amountSupplied,
-        amountBorrowed,
-        supplyTokenPrice,
-        borrowTokenPrice,
-        duration,
-    });
-
-    return (
-        <Card>
-            <CardHeader className='pb-[12px]'>
-                <div className="flex justify-between items-center gap-[12px]">
-                    <BodyText level='body2' weight='normal' className="text-gray-600">
-                        Estimate returns by using slider below
-                    </BodyText>
-                    <div className="flex items-center gap-[8px]">
-                        <BodyText level='body2' weight='normal' className="text-gray-600">
-                            Your earnings
-                        </BodyText>
-                        <HeadingText level='h4' weight='medium' className="text-gray-800">
-                            ${abbreviateNumber(estimatedEarnings)}
-                        </HeadingText>
-                    </div>
-                </div>
-            </CardHeader>
-            <CardContent className='bg-white rounded-5 px-[32px] py-[28px]'>
-                <div className="flex flex-col gap-[16px]">
-                    {
-                        rows.map(row => (
-                            <div key={row.id} className="flex flex-col gap-[16px]">
-                                <div className="flex items-center justify-between">
-                                    <div className="flex items-center gap-[8px]">
-                                        <BodyText level='body2' weight='normal' className="capitalize text-gray-600">
-                                            {row.title} -
-                                        </BodyText>
-                                        <div className="flex items-center gap-[6px]">
-                                            {row.logo && <ImageWithDefault src={row.logo} alt={row.selectedLabel} width={20} height={20} className='rounded-full max-w-[20px] max-h-[20px]' />}
-                                            <HeadingText level='h4' weight='medium' className="text-gray-800">
-                                                {row.selectedValue} {row.selectedLabel}
-                                            </HeadingText>
-                                        </div>
-                                    </div>
-                                    <BodyText level='body1' weight='normal' className="text-gray-600">
-                                        {row.totalValue}
-                                    </BodyText>
-                                </div>
-                                <Slider defaultValue={[row.selectedValue]} max={row.totalValue} step={row.step} onValueChange={(value) => handleSelectedValueChange(value[0], row.key)} />
-                            </div>
-                        ))
-                    }
-                </div>
-            </CardContent>
-        </Card>
-
-    )
-}
-
-// Helper functions
-function getStatDisplayValue(value: string | number, hasPrefix: boolean = true) {
-    return `${hasPrefix ? hasLowestDisplayValuePrefix(Number(value)) : ""}${getLowestDisplayValue(Number(value))}`;
-}
-
-// Function to calculate estimated earnings
-/**
- * @param lendCollateral - Amount of collateral supplied
- * @param borrowing - Amount of borrowing
- * @param duration - Duration of investment
- * @returns - Estimated earnings
- */
-
-/** Calculation logic - 
-Supply APY: 
-𝑅𝑠 (as a decimal, e.g., 0.05 for 5%)
-
-Borrow APY: 
-𝑅𝑏 (as a decimal, e.g., 0.04 for 4%)
-
-Amount supplied: 𝐴𝑠
-Amount borrowed: 𝐴𝑏
-
-supply token price: Price(S)
-borrow token price: Price(B)
-
-Duration of investment: 𝑇
-T (in years; for shorter periods, express as a fraction, e.g., 0.5 for 6 months)
-
-Net Returns = T×(As × Rs x Price(S) ​− Ab ​× Rb x Price(B))
-*/
-function getEstimatedEarnings({
-    supplyAPY,
-    borrowAPY,
-    amountSupplied,
-    amountBorrowed,
-    supplyTokenPrice,
-    borrowTokenPrice,
-    duration,
-}: {
-    supplyAPY: number;
-    borrowAPY: number;
-    amountSupplied: number;
-    amountBorrowed: number;
-    supplyTokenPrice: number;
-    borrowTokenPrice: number;
-    duration: number;
-}) {
-    return duration * (amountSupplied * supplyAPY * supplyTokenPrice - amountBorrowed * borrowAPY * borrowTokenPrice);
 }
